refactor(music-player): move MusbxApi to it's own folder
Since it is no longer just related to the Demixer
