+++ lib/music_player/musbx_api/exceptions.dart
